+++ netutil/netutil_darwin_test.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package netutil
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -11,9 +8,7 @@ package netutil
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"io/ioutil"
 	"os"
-	"runtime"
 	"testing"
 
 	. "pkg.re/essentialkaos/check.v1"
@@ -31,20 +26,6 @@ var _ = Suite(&NetUtilSuite{})
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func (s *NetUtilSuite) TestCommon(c *C) {
-	if runtime.GOOS == "linux" {
-		c.Assert(getDefaultRouteInterface(), Not(Equals), "")
-
-		procRouteFile = "/tmp/not-exist"
-
-		c.Assert(getDefaultRouteInterface(), Equals, "")
-
-		procRouteFile = s.CreateTestFile(c, "Iface   Destination     Gateway         Flags   RefCnt  Use     Metric  Mask            MTU     Window  IRTT\neth0    0070652E        00000000        0001    0       0       0       00F0FFFF        0       0       0")
-
-		c.Assert(getDefaultRouteInterface(), Equals, "")
-	}
-}
-
 func (s *NetUtilSuite) TestGetIP(c *C) {
 	c.Assert(GetIP(), Not(Equals), "")
 }
@@ -53,17 +34,4 @@ func (s *NetUtilSuite) TestGetIP6(c *C) {
 	if os.Getenv("CI") == "" {
 		c.Assert(GetIP6(), Not(Equals), "")
 	}
-}
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-func (s *NetUtilSuite) CreateTestFile(c *C, data string) string {
-	tmpDir := c.MkDir()
-	tmpFile := tmpDir + "/test.file"
-
-	if ioutil.WriteFile(tmpFile, []byte(data), 0644) != nil {
-		c.Fatal("Can't create temporary file")
-	}
-
-	return tmpFile
 }
